requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
black==23.9.1
click==8.1.7
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
promise==2.3
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
